fix distance formula

+++ include/Graph/Graph.hpp
@@ -24,7 +24,7 @@ class Graph{
         std::vector<int> UnitaryToOriginalDemandID;
 
         //warning, the adjacency matrix keep -1 if m_vertices aren't adjacent and the length of it's edge otherwise
-        std::vector<std::vector<int>> adjacencyMatrix;
+        std::vector<std::vector<double>> adjacencyMatrix;
 
         std::vector<std::vector<double>> TSPKernelTime;
         std::vector<std::vector<std::vector<int>>> TSPKernelPath;
@@ -49,8 +49,8 @@ class Graph{
         const std::vector<std::vector<Edge>>& getAdjacencyList() const;
         const std::vector<Edge>& getAdjacencyList(int vertexID) const;
 
-        const std::vector<std::vector<int>>& getAdjacencyMatrix() const;
-        const std::vector<int>& getAdjacencyMatrix(int vertexID) const;
+        const std::vector<std::vector<double>>& getAdjacencyMatrix() const;
+        const std::vector<double>& getAdjacencyMatrix(int vertexID) const;
 
         const std::vector<Vertex>& getVertices() const;
         const Vertex& getVertice(int GraphID) const;
